% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantec_get_data.R
\name{quantec_get_data}
\alias{quantec_get_data}
\title{Get Quantec Data}
\usage{
quantec_get_data(
  time_series_code,
  freq = c("M", "Q", "A")[1],
  start_year = "1900-01-01",
  end_year = Sys.Date(),
  respformat = c("csv", "json")[1],
  log_file,
  is_tidy = TRUE
)
}
\arguments{
\item{time_series_code}{time series code to return, \verb{NMS-EC_BUS,NMS-GA_BUS}}

\item{freq}{frequency to return \code{M}, \code{Q} or \code{A}}

\item{start_year}{year to start}

\item{end_year}{year to end}

\item{respformat}{to return \code{csv} or \code{json}}

\item{log_file}{log file to output to if in parallel}

\item{is_tidy}{tidyformat for easy read}
}
\value{
tibble
}
\description{
Get timeseries by specifying comma seperated code(s).
}
